Split out Image Constructing builtins...
Move remaining under Images Misc to be split out even more, later.
