+++ mathics/builtin/image/__init__.py
@@ -1,3 +1,6 @@
 """
 Image Manipulation
+
+
+For the full compliment of functions, you need to have scikit-image installed.
 """
